Update 2016/Day25/day25.c
The safety_counter is used to prevent infinite loops and is limited to 10000. On the 6502 architecture, 32-bit long operations are significantly more expensive than 16-bit int operations. Since 10000 fits within a 16-bit int, changing the type here can provide a noticeable performance improvement.

+++ 2016/Day25/day25.c
@@ -82,7 +82,7 @@ int run_with_a(long initial_a) {
     int pc = 0;
     int next_expected = 0;
     int outputs_found = 0;
-    long safety_counter = 0;
+    int safety_counter = 0;
     Instruction* instr;
     
     regs[0] = initial_a;
